pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sprite-splitter"
version = "1.0.0"
description = "A desktop utility to detect, organize, name, and export individual sprite frames from sprite sheet PNGs."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Sprite Splitter Contributors"},
]
dependencies = [
    "PySide6>=6.6,<7",
    "Pillow>=10.0,<13",
    "numpy>=1.24,<3",
    "opencv-python-headless>=4.8,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
]

[project.scripts]
sprite-splitter = "sprite_splitter.main:main"

[project.gui-scripts]
sprite-splitter-gui = "sprite_splitter.main:main"

[tool.setuptools.packages.find]
where = ["src"]
